More calculation can be performed after display result
Before this when we try to make calculation after display result like +5 with
result 43 can't be done.

Calculator need to perform calculation after the result is shown.

The problem is in the resetAll() function.By calling it clear the total
variable and set it to null.So we remove the code total = null from
the resetAll function.Thus the problem is solved.